requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pydantic
PyPDF2
openai
